ENH: create folder for mock
